pyproject.toml
[tool.poetry]
name = "hack"
version = "0.1.0"
description = ""
authors = ["Tim <jlemyp96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
djangorestframework_simplejwt = "^4.3"
django-extensions = "^2.2"
Werkzeug = "^0.15.5"
django-reversion = "^3.0"
ipython = "^7.7"
django-filter = "^2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
